+++ import_export_extensions/admin/forms/__init__.py
@@ -1,2 +1,3 @@
-from .export_admin_form import ExportJobAdminForm
-from .import_admin_form import ImportJobAdminForm
+from .export_job_admin_form import ExportJobAdminForm
+from .import_admin_form import ForceImportForm
+from .import_job_admin_form import ImportJobAdminForm