+++ import_export_extensions/api/serializers/import_job.py
@@ -72,6 +72,7 @@ class ImportJobSerializer(serializers.ModelSerializer):
             "input_errors_file",
             "import_started",
             "import_finished",
+            "force_import",
             "created",
             "modified",
         )
@@ -87,10 +88,8 @@ class CreateImportJob(serializers.Serializer):
     resource_class: typing.Type[resources.CeleryModelResource]
 
     file = serializers.FileField(required=True)
-    skip_parse_step = serializers.BooleanField(
-        required=False,
-        default=False,
-    )
+    force_import = serializers.BooleanField(default=False, required=False)
+    skip_parse_step = serializers.BooleanField(default=False, required=False)
 
     def __init__(
         self,
@@ -111,10 +110,11 @@ class CreateImportJob(serializers.Serializer):
         """Create import job."""
         return models.ImportJob.objects.create(
             data_file=validated_data["file"],
+            force_import=validated_data["force_import"],
+            skip_parse_step=validated_data["skip_parse_step"],
             resource_path=self.resource_class.class_path,
             resource_kwargs=self._resource_kwargs,
             created_by=self._user,
-            skip_parse_step=validated_data["skip_parse_step"],
         )
 
     def update(self, instance, validated_data):