+++ tests/conftest.py
@@ -1,6 +1,9 @@
 from django.conf import settings
 from django.contrib.auth import get_user_model
+from django.contrib.auth.models import User
 from django.core.files.uploadedfile import SimpleUploadedFile
+
+from rest_framework import test
 
 import pytest
 
@@ -61,9 +64,34 @@ def membership(band: Band) -> Membership:
 
 
 @pytest.fixture
-def uploaded_file(existing_artist: Artist):
+def uploaded_file(existing_artist: Artist) -> SimpleUploadedFile:
     """Generate valid `Artist` import file."""
     import_job = ArtistImportJobFactory.build(artists=[existing_artist])
+    return SimpleUploadedFile(
+        "test_file.csv",
+        content=import_job.data_file.file.read().encode(),
+        content_type="text/plain",
+    )
+
+
+@pytest.fixture
+def force_import_artist_job(new_artist: Artist) -> Artist:
+    """Return `ImportJob` with `force_import=True` and file with invalid row."""
+    return ArtistImportJobFactory(
+        artists=[new_artist],
+        is_valid_file=False,
+        force_import=True,
+    )
+
+
+@pytest.fixture
+def invalid_uploaded_file(new_artist: Artist) -> SimpleUploadedFile:
+    """Generate invalid `Artist` imort file."""
+    import_job = ArtistImportJobFactory.build(
+        artists=[new_artist],
+        force_import=True,
+        is_valid_file=False,
+    )
     return SimpleUploadedFile(
         "test_file.csv",
         content=import_job.data_file.file.read().encode(),
@@ -102,3 +130,19 @@ def temp_directory_for_media(tmpdir_factory):
         )
     media = tmpdir_factory.mktemp("tmp_media")
     settings.MEDIA_ROOT = media
+
+
+@pytest.fixture
+def api_client() -> test.APIClient:
+    """Create api client."""
+    return test.APIClient()
+
+
+@pytest.fixture
+def admin_api_client(
+    superuser: User,
+    api_client: test.APIClient,
+) -> test.APIClient:
+    """Authenticate admin_user and return api client."""
+    api_client.force_authenticate(user=superuser)
+    return api_client