+++ tests/settings.py
@@ -9,6 +9,8 @@ BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 # SECURITY WARNING: keep the secret key used in production secret!
 SECRET_KEY = "a87082n4v52u4rnvk2edv128eudfvn5"
 
+ALLOWED_HOSTS = ["*"]
+
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True